Deleting files

+++ controllers/admin.js
@@ -2,6 +2,7 @@ const Product = require('../models/product')
 const validationResult = require('express-validator')
 
 const mongoose = require('mongoose')
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
     /*if (!req.session.isLoggedIn) {
@@ -171,6 +172,7 @@ exports.postEditProduct = (req, res, next) => {
             product.description = updatedDescription
 
             if (image) {
+                fileHelper.deleteFile(product.imageUrl)
                 product.imageUrl = image.path
             }
             
@@ -210,15 +212,22 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-        .then(result => {
-            console.log('Deleted Product!')
-            res.redirect('/admin/products')
-        })
-        .catch(err => {
-            console.error(err)
-            const error = new Error(err)
-            error.httpStatusCode = 500
-            return next(error)
-        })
+    Product.findById(prodId)
+            .then(product => {
+                if (!product) {
+                    return next(new Error('Product not found.'))
+                }
+                fileHelper.deleteFile(product.imageUrl)
+                return  Product.deleteOne({_id: prodId, userId: req.user._id})
+            })
+            .then(result => {
+                console.log('Deleted Product!')
+                res.redirect('/admin/products')
+            })
+            .catch(err => {
+                console.error(err)
+                const error = new Error(err)
+                error.httpStatusCode = 500
+                return next(error)
+            })
 }
